プロジェクトの構造を変更
