--- app/src/main/java/jp/androidbook/recyclerviewtemplate/simple/SimpleText.kt
@@ -1,5 +1,0 @@
-package jp.androidbook.recyclerviewtemplate.simple
-
-data class SimpleText(
-    val text: String
-)
