+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/ui/LiveEvent.kt
@@ -1,4 +1,4 @@
-package jp.androidbook.recyclerviewtemplate
+package jp.androidbook.recyclerviewtemplate.ui
 
 import androidx.annotation.MainThread
 import androidx.lifecycle.LifecycleOwner
