+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/ui/MainModule.kt
@@ -1,4 +1,4 @@
-package jp.androidbook.recyclerviewtemplate.di
+package jp.androidbook.recyclerviewtemplate.ui
 
 import dagger.Module
 import dagger.hilt.InstallIn
