+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/ui/OnTappedRecyclerView.kt
@@ -1,4 +1,4 @@
-package jp.androidbook.recyclerviewtemplate
+package jp.androidbook.recyclerviewtemplate.ui
 
 interface OnTappedRecyclerViewListener {
     fun onTapped(text: String)
