+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/ui/simple/SimpleListViewHolder.kt
@@ -1,27 +1,25 @@
-package jp.androidbook.recyclerviewtemplate.update
+package jp.androidbook.recyclerviewtemplate.ui.simple
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import jp.androidbook.recyclerviewtemplate.R
 import jp.androidbook.recyclerviewtemplate.databinding.ItemSimpleRecyclerViewBinding
 
-class UpdateViewHolder(
-    binding: ItemSimpleRecyclerViewBinding
+class SimpleListViewHolder(
+    private val binding: ItemSimpleRecyclerViewBinding
 ) : RecyclerView.ViewHolder(binding.root) {
-    private val textView: TextView = itemView.findViewById(R.id.textView)
 
-    fun setText(text: String) {
-        textView.text = text
+    fun bind(simpleText: SimpleText, viewModel: SimpleListViewModel) {
+        binding.simpleText = simpleText
+        binding.viewModel = viewModel
     }
 
     companion object {
-        fun createViewHolder(parent: ViewGroup): UpdateViewHolder {
+        fun createViewHolder(parent: ViewGroup): SimpleListViewHolder {
             val layoutInflater = LayoutInflater.from(parent.context)
             val binding = ItemSimpleRecyclerViewBinding.inflate(layoutInflater, parent, false)
 
-            return UpdateViewHolder(binding)
+            return SimpleListViewHolder(binding)
         }
     }
 }
