+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/ui/simple/SimpleText.kt
@@ -1,0 +1,5 @@
+package jp.androidbook.recyclerviewtemplate.ui.simple
+
+data class SimpleText(
+    val text: String
+)
