+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/ui/update/UpdateViewHolder.kt
@@ -1,25 +1,27 @@
-package jp.androidbook.recyclerviewtemplate.simple
+package jp.androidbook.recyclerviewtemplate.ui.update
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import jp.androidbook.recyclerviewtemplate.R
 import jp.androidbook.recyclerviewtemplate.databinding.ItemSimpleRecyclerViewBinding
 
-class SimpleListViewHolder(
-    private val binding: ItemSimpleRecyclerViewBinding
+class UpdateViewHolder(
+    binding: ItemSimpleRecyclerViewBinding
 ) : RecyclerView.ViewHolder(binding.root) {
+    private val textView: TextView = itemView.findViewById(R.id.textView)
 
-    fun bind(simpleText: SimpleText, viewModel: SimpleListViewModel) {
-        binding.simpleText = simpleText
-        binding.viewModel = viewModel
+    fun setText(text: String) {
+        textView.text = text
     }
 
     companion object {
-        fun createViewHolder(parent: ViewGroup): SimpleListViewHolder {
+        fun createViewHolder(parent: ViewGroup): UpdateViewHolder {
             val layoutInflater = LayoutInflater.from(parent.context)
             val binding = ItemSimpleRecyclerViewBinding.inflate(layoutInflater, parent, false)
 
-            return SimpleListViewHolder(binding)
+            return UpdateViewHolder(binding)
         }
     }
 }
