+++ app/src/main/res/layout/fragment_grid_selection.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".grid.GridSelectionFragment">
+        tools:context=".ui.grid.GridSelectionFragment">
 
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/gridSelectionTabLayout"
