+++ app/src/main/res/layout/fragment_simple_recycler.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".simple.SimpleRecyclerFragment">
+        tools:context=".ui.simple.SimpleRecyclerFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/simpleRecyclerView"
