+++ app/src/main/res/layout/fragment_staggered_grid.xml
@@ -6,7 +6,7 @@
    <androidx.constraintlayout.widget.ConstraintLayout
        android:layout_width="match_parent"
        android:layout_height="match_parent"
-       tools:context=".staggered.StaggeredGridFragment">
+       tools:context=".ui.staggered.StaggeredGridFragment">
 
       <androidx.recyclerview.widget.RecyclerView
           android:id="@+id/staggeredRecyclerView"
