+++ app/src/main/res/layout/fragment_three_grid.xml
@@ -5,7 +5,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".grid.three.ThreeGridFragment">
+        tools:context=".ui.grid.three.ThreeGridFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/threeGridRecyclerView"
