+++ app/src/main/res/layout/fragment_two_grid.xml
@@ -5,7 +5,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".grid.two.TwoGridFragment">
+        tools:context=".ui.grid.two.TwoGridFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/twoGridRecyclerView"
