+++ app/src/main/res/layout/fragment_update.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".update.UpdateFragment">
+        tools:context=".ui.update.UpdateFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/updateRecyclerView"
